Alterações em Services e finalizado com correção de erros

+++ dentalclinic/src/main/java/br/com/dentalclinic/service/impl/UsuarioServiceImpl.java
@@ -1,9 +1,11 @@
 package br.com.dentalclinic.service.impl;
 
 import br.com.dentalclinic.dto.ClinicaDTO;
+import br.com.dentalclinic.dto.PacienteDTO;
 import br.com.dentalclinic.dto.TipoUsuarioDTO;
 import br.com.dentalclinic.dto.UsuarioDTO;
 import br.com.dentalclinic.model.Clinica;
+import br.com.dentalclinic.model.Paciente;
 import br.com.dentalclinic.model.TipoUsuario;
 import br.com.dentalclinic.model.Usuario;
 import br.com.dentalclinic.repository.IUsuarioRepository;
@@ -12,6 +14,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -41,20 +44,37 @@ public class UsuarioServiceImpl implements IService<UsuarioDTO> {
         return usuarioDTO;
     }
 
+    @Override
     public Optional<UsuarioDTO> buscarById(Integer id) {
-        return usuarioRepository.findById(id);
+        Usuario usuario = usuarioRepository.findById(id).get();
+        UsuarioDTO usuarioDTO = mapperEntityToDTO(usuario);
+        return Optional.ofNullable(usuarioDTO);
     }
 
     public Optional<Usuario> buscarByEmail(String email){return usuarioRepository.findByEmail(email);}
+    @Override
+    public List<UsuarioDTO> buscarTodos(){
+        List<Usuario> usuarios = usuarioRepository.findAll();
+        List<UsuarioDTO> usuarioDTOS = new ArrayList<>();
 
-    public List<UsuarioDTO> buscarTodos(){return usuarioRepository.findAll();}
-
-    public Usuario atualizar(Usuario usuario) {
-        return usuarioRepository.saveAndFlush(usuario);
+        for (Usuario usuario : usuarios){
+            UsuarioDTO usuarioDTO = mapperEntityToDTO(usuario);
+            usuarioDTOS.add(usuarioDTO);
+        }
+        return usuarioDTOS;
     }
-
+    @Override
+    public UsuarioDTO atualizar(UsuarioDTO usuarioDTO) {
+        Usuario usuario = mapperDTOToEntity(usuarioDTO);
+        usuario = usuarioRepository.saveAndFlush(usuario);
+        usuarioDTO = mapperEntityToDTO(usuario);
+        return usuarioDTO;
+    }
+    @Override
     public void deletar(Integer id) {
-        usuarioRepository.deleteById(id);
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+        }
     }
 
     public boolean ifUsuarioExists(int id){
